Create the new directory structrue under "New Site" directory.
This is just to keep old site while developing the new one. When new site is
advanced enough, New Site will change to be the root and config.yaml may return
to its former configuration (maximun of 3 nesting levels).

All pages are empty. Next task is moving information from the old site to its
place in the new site.